add yapapi.payload deprecate yapapi.package
+ add "Deprecated" package to enable controlled deprecation of features
