+++ pyproject.toml
@@ -45,6 +45,7 @@ colorama = "^0.4.4"
 # Note that putting `goth` in `poetry.dev-dependencies` instead of `poetry.dependencies`
 # would not work: see https://github.com/python-poetry/poetry/issues/129.
 goth = { version = "^0.2.1", optional = true, python = "^3.8.0" }
+Deprecated = "^1.2.12"
 
 [tool.poetry.extras]
 cli = ['fire', 'rich']
