+++ yapapi/props/inf.py
@@ -2,6 +2,7 @@
 
 from typing import Optional, List
 from dataclasses import dataclass, field
+from deprecated import deprecated
 from enum import Enum
 from .base import Model
 
@@ -10,13 +11,17 @@ INF_STORAGE: str = "golem.inf.storage.gib"
 INF_CORES: str = "golem.inf.cpu.cores"
 TRANSFER_CAPS: str = "golem.activity.caps.transfer.protocol"
 
+RUNTIME_WASMTIME = "wasmtime"
+RUNTIME_EMSCRIPTEN = "emscripten"
+RUNTIME_VM = "vm"
 
 @dataclass
+@deprecated(version="0.6.0", reason="please use yapapi.props.inf.RUNTIME_* constants directly", action="default")
 class RuntimeType(Enum):
     UNKNOWN = ""
-    WASMTIME = "wasmtime"
-    EMSCRIPTEN = "emscripten"
-    VM = "vm"
+    WASMTIME = RUNTIME_WASMTIME
+    EMSCRIPTEN = RUNTIME_EMSCRIPTEN
+    VM = RUNTIME_VM
 
 
 @dataclass
@@ -27,16 +32,16 @@ class WasmInterface(Enum):
 
 @dataclass
 class InfBase(Model):
-    mem: float = field(metadata={"key": INF_MEM})
-    runtime: RuntimeType = field(metadata={"key": "golem.runtime.name"})
+    runtime: str = field(metadata={"key": "golem.runtime.name"})
 
+    mem: float = field(metadata={"key": INF_MEM})
     storage: Optional[float] = field(default=None, metadata={"key": INF_STORAGE})
     transfers: Optional[List[str]] = field(default=None, metadata={"key": TRANSFER_CAPS})
 
 
 @dataclass
 class InfVm(InfBase):
-    runtime = RuntimeType.VM
+    runtime = RUNTIME_VM
     cores: int = field(default=1, metadata={"key": INF_CORES})
